Add generating poster in new tab
