+++ web/src/app/import-export/poster/poster.component.html
@@ -23,7 +23,10 @@
         </tbody>
       </table>
 
-      <button class="btn btn-primary" (click)="refreshPicture()">Refresh picture</button>
+      <div class="row">
+        <button class="btn btn-primary mx-2" (click)="refreshPicture()">Refresh picture</button>
+        <button class="btn btn-primary mx-2" (click)="generatePoster()">Generate poster</button>
+      </div>
     </div>
 
     <div class="col-lg-8">
